Switch from Webpack to Parcel
Webpack was overly complicated for this project, and even with plenty of
configuration, it was still performing sub-optimally. Instead of
configuring further, this switches the bundler to Parcel, which is
sufficient for such a simple app.

This also includes a slight restructure, to move the static assets into
src so dist can remain a build-only directory.